Introduce serial interface to output uVisor's log message
Currently only LED blinking and semihosting can be used to represent
the status of uVisor; however neither of them is convenient in debugging.
For efficient debugging, Serial is a good solution and its interface
is defined first in this commit.

The detail implementations of serial interface is still empty and
will be fulfilled in the future.